+++ src/main/scala/com/databricks/labs/mosaic/expressions/raster/base/RasterExpressionSerialization.scala
@@ -34,12 +34,13 @@ trait RasterExpressionSerialization {
         expressionConfig: MosaicExpressionConfig
     ): Any = {
         if (returnsRaster) {
+            val tile = data.asInstanceOf[MosaicRasterTile]
             val checkpoint = expressionConfig.getRasterCheckpoint
             val rasterType = outputDataType.asInstanceOf[StructType].fields(1).dataType
-            val result = data
-                .asInstanceOf[MosaicRasterTile]
+            val result = tile
                 .formatCellId(IndexSystemFactory.getIndexSystem(expressionConfig.getIndexSystem))
                 .serialize(rasterType, checkpoint)
+            RasterCleaner.dispose(tile)
             result
         } else {
             data