+++ src/main/scala/com/databricks/labs/mosaic/expressions/raster/base/RasterGeneratorExpression.scala
@@ -72,13 +72,13 @@ abstract class RasterGeneratorExpression[T <: Expression: ClassTag](
 
     override def eval(input: InternalRow): TraversableOnce[InternalRow] = {
         GDAL.enable(expressionConfig)
-        val generatedRasters =
-            rasterGenerator(MosaicRasterTile.deserialize(rasterExpr.eval(input).asInstanceOf[InternalRow], cellIdDataType))
+        val tile = MosaicRasterTile.deserialize(rasterExpr.eval(input).asInstanceOf[InternalRow], cellIdDataType)
+        val generatedRasters = rasterGenerator(tile)
 
         // Writing rasters disposes of the written raster
         val rows = generatedRasters.map(_.formatCellId(indexSystem).serialize())
         generatedRasters.foreach(gr => RasterCleaner.dispose(gr))
-        GDAL.dropDrivers()
+        RasterCleaner.dispose(tile)
 
         rows.map(row => InternalRow.fromSeq(Seq(row)))
     }