+++ src/main/scala/com/databricks/labs/mosaic/functions/MosaicExpressionConfig.scala
@@ -2,9 +2,8 @@ package com.databricks.labs.mosaic.functions
 
 import com.databricks.labs.mosaic._
 import com.databricks.labs.mosaic.core.index.IndexSystemFactory
-import org.apache.spark.SparkConf
-import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.types.DataType
+import org.apache.spark.sql.{RuntimeConfig, SparkSession}
 
 /**
   * Mosaic Expression Config is a class that contains the configuration for the
@@ -35,8 +34,8 @@ case class MosaicExpressionConfig(configs: Map[String, String]) {
 
     def getCellIdType: DataType = IndexSystemFactory.getIndexSystem(getIndexSystem).cellIdType
 
-    def setGDALConf(conf: SparkConf): MosaicExpressionConfig = {
-        val toAdd = conf.getAllWithPrefix(MOSAIC_GDAL_PREFIX)
+    def setGDALConf(conf: RuntimeConfig): MosaicExpressionConfig = {
+        val toAdd = conf.getAll.filter(_._1.startsWith(MOSAIC_GDAL_PREFIX))
         MosaicExpressionConfig(configs ++ toAdd)
     }
 
@@ -74,7 +73,7 @@ object MosaicExpressionConfig {
             .setGeometryAPI(spark.conf.get(MOSAIC_GEOMETRY_API, JTS.name))
             .setIndexSystem(spark.conf.get(MOSAIC_INDEX_SYSTEM, H3.name))
             .setRasterCheckpoint(spark.conf.get(MOSAIC_RASTER_CHECKPOINT, MOSAIC_RASTER_CHECKPOINT_DEFAULT))
-            .setGDALConf(spark.sparkContext.getConf)
+            .setGDALConf(spark.conf)
 
     }